Initial re-implementation of player spells and research

+++ src/toniarts/openkeeper/game/controller/IObjectsController.java
@@ -36,7 +36,7 @@ public interface IObjectsController extends IEntityWrapper<IObjectController> {
 
     public EntityId loadObject(short objectId, short ownerId, Vector3f pos, float rotation);
 
-    public EntityId loadObject(short objectId, short ownerId, int x, int y, Integer money, Integer spellId);
+    public EntityId loadObject(short objectId, short ownerId, int x, int y, Integer money, Short spellId);
 
     public EntityId addRoomGold(short ownerId, int x, int y, int money, int maxMoney);
